+++ documents/prototypage/src/Minus.java
@@ -11,7 +11,7 @@ public class Minus extends Operator
 	}
 
 	@Override
-	public float evaluate() throws Exception
+	public double evaluate() throws Exception
 	{
 		try {
 			return this.expLeft.evaluate() - this.expRight.evaluate();
