+++ documents/prototypage/src/Pow.java
@@ -12,7 +12,7 @@ public class Pow extends Operator
 	}
 
     @Override
-    public float evaluate() throws Exception {
+    public double evaluate() throws Exception {
         try {
 			return (float)Math.pow(this.expLeft.evaluate(), this.expRight.evaluate());
 		} catch (VariableEvaluationException e) {
