+++ documents/prototypage/src/Times.java
@@ -13,7 +13,7 @@ public class Times extends Operator
 	}
 
 	@Override
-	public float evaluate() throws Exception
+	public double evaluate() throws Exception
 	{
 		try {
 			return this.expLeft.evaluate() * this.expRight.evaluate();
